% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{AddGOGeneList.manual}
\alias{AddGOGeneList.manual}
\title{AddGOGeneList.manual}
\usage{
AddGOGeneList.manual(
  obj = combined.obj,
  GO = "GO:0034976",
  web.open = F,
  genes = c("A0A140VKG3", "ARX", "CNTN2", "DRD1", "DRD2", "FEZF2", "LHX6")
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{GO}{GO-term; Default: 'GO:0034976'}

\item{web.open}{Open weblink, Default: F}

\item{genes}{Charcter vector of genes, Default: c("A0A140VKG3", "ARX", "CNTN2", "DRD1", "DRD2", "FEZF2", "LHX6")}
}
\description{
Add a GO-term gene list under obj@misc$GO$xxxx.
}
\seealso{
\code{\link[Stringendo]{iprint}}
}
