% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{AddGOScore}
\alias{AddGOScore}
\title{AddGOScore}
\usage{
AddGOScore(obj = combined.obj, GO = "GO:0034976", FixName = TRUE)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{GO}{GO Term. String of form "GO:xxxxxxx", Default: 'GO:0034976'.}

\item{FixName}{Shall trailing "1" ar the end of the new column name be removed?}
}
\description{
Add GO Score to a Seurat object
}
\seealso{
\code{\link[Seurat]{AddModuleScore}}
}
