% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{aut.res.clustering}
\alias{aut.res.clustering}
\title{aut.res.clustering}
\usage{
aut.res.clustering(
  obj = combined.obj,
  lower.median = 100,
  upper.median = 200,
  lower.res = round(ncol(obj)/10000),
  upper.res = round(ncol(obj)/1000),
  assay = c("integrated", "RNA")[1],
  max.loop = 20
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{lower.median}{Lower limit of median cell count per cluster, Default: 100}

\item{upper.median}{Upper limit of median cell count per cluster, Default: 200}

\item{lower.res}{Lower limit of clustering resolution search, Default: round(ncol(obj)/10000)}

\item{upper.res}{Upper limit of clustering resolution search, Default: round(ncol(obj)/1000)}

\item{assay}{Which assay to use? Default: Seurat::DefaultAssay(obj)}

\item{max.loop}{Maximum iteration count, Default: 20}
}
\description{
Find optimal granule (clustering) resolution for sterss classification
}
\seealso{
\code{\link[Seurat]{reexports}}, \code{\link[Seurat]{FindClusters}}
\code{\link[Stringendo]{iprint}}
}
