% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{Auto.GO.thresh}
\alias{Auto.GO.thresh}
\title{Automated GO Term-based Thresholding without using a Shiny app.}
\usage{
Auto.GO.thresh(
  obj = combined.obj,
  proposed.method = c("fitted", "empirical")[1],
  quantile = c(0.99, 0.9)[1],
  stress.ident1,
  stress.ident2,
  notstress.ident3,
  notstress.ident4 = NULL,
  plot.results = TRUE,
  ...
)
}
\arguments{
\item{obj}{Seurat single cell object, with Granule Averages GO-Scores in metadata (ident1-4).}

\item{proposed.method}{The method for estimating thresholds, default: \code{"fitted"}.
Options are \code{"fitted"} or \code{"empirical"}.}

\item{quantile}{The quantile cutoff for threshold determination, default: \code{0.99}.
Options are \code{0.99} or \code{0.9}.}

\item{stress.ident1}{Identifier for the first stress-related GO term.}

\item{stress.ident2}{Identifier for the second stress-related GO term.}

\item{notstress.ident3}{Identifier for the first non-stress-related GO term. Idents 3 and 4 act
as a negative filter for stress identification.}

\item{notstress.ident4}{Identifier for the second non-stress-related GO term, optional.}

\item{plot.results}{Boolean flag to control the plotting of results on UMAPs and histograms,
default: \code{TRUE}.}
}
\description{
Automatically applies granular GO term-based thresholding to identify stressed cells
in a Seurat object without launching a Shiny app. This function calculates the
"granule average GO-scores" for individual GO terms and updates the object with a combined stress
identification based on these thresholds.
}
