% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{calc.cluster.averages.gruffi}
\alias{calc.cluster.averages.gruffi}
\title{calc.cluster.averages.gruffi}
\usage{
calc.cluster.averages.gruffi(
  col_name = "Score.GO.0006096",
  obj = combined.obj,
  split_by = Seurat.utils::GetClusteringRuns(obj)[1],
  stat = c("mean", "median", "normalized.mean", "normalized.median")[3],
  scale.zscore = FALSE,
  quantile.thr = 0.99,
  absolute.thr = FALSE,
  plotit = FALSE,
  save.plot = FALSE,
  return.plot = FALSE,
  max.bin.plot = 200,
  simplify = TRUE,
  suffix = NULL,
  ylab.text = paste("Cluster", stat, "score"),
  title = paste("Cluster", stat, col_name),
  subtitle = NULL,
  width = 8,
  height = 6,
  xlb = if (absolute.thr) {
     paste("Threshold at", absolute.thr)
 } else {
    
    paste("Black lines: ",
    Stringendo::kppd(Stringendo::percentage_formatter(quantile.thr)), "quantiles |",
    "Cl. >", Stringendo::percentage_formatter(quantile.thr), "are highlighted. |",
    split_by)
 },
  ...
)
}
\arguments{
\item{col_name}{Numeric metadata column name, Default: 'Score.GO.0006096'}

\item{obj}{Seurat single cell object, Default: combined.obj}

\item{split_by}{Clustering to split by (categorical metadata column name), Default: Seurat.utils::GetClusteringRuns(obj)\link{1}}

\item{stat}{How to caluclate the central tendency? Default: c("mean", "median", "normalized.mean", "normalized.median")\link{3}}

\item{scale.zscore}{Scale as zscore?, Default: FALSE}

\item{quantile.thr}{Quantile threshold for cutoff, Default: 0.99}

\item{absolute.thr}{Use an absolute threshold? Default: FALSE}

\item{plotit}{Draw a plot? Default: FALSE}

\item{save.plot}{Save plot into a file. Default: FALSE}

\item{return.plot}{Return the plot as output of the function? Default: FALSE}

\item{max.bin.plot}{What is considered too many clusters for a histogram to plot? Default: 200}

\item{simplify}{Simplify output? Default: TRUE}

\item{suffix}{What suffix to append at the end of the plotname/file name? Default: NULL}

\item{ylab.text}{Y-axis label. Default: paste("Cluster", stat, "score")}

\item{title}{Plot title, Default: paste("Cluster", stat, col_name)}

\item{subtitle}{Plot subtitle, Default: NULL}

\item{width}{Plot width, Default: 8}

\item{height}{Plot height, Default: 6}

\item{xlb}{X-axis label, Default: if (absolute.thr) paste("Threshold at", absolute.thr) else paste("Black lines: ",
Stringendo::kppd(Stringendo::percentage_formatter(quantile.thr)), "quantiles |",
"Cl. >", Stringendo::percentage_formatter(quantile.thr),
"are highlighted. |", split_by)}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Calculate granule (cluster) averages scores.
}
\seealso{
\code{\link[Stringendo]{percentage_formatter}}, \code{\link[Stringendo]{iprint}}
\code{\link[dplyr]{select_all}}, \code{\link[dplyr]{group_by_all}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
\code{\link[rlang]{sym}}
\code{\link[CodeAndRoll2]{sem}}, \code{\link[CodeAndRoll2]{sortbyitsnames}}
}
