% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{CalcClusterAverages_Gruffi}
\alias{CalcClusterAverages_Gruffi}
\title{Calculate Cluster Averages for Granules in a Seurat Object}
\usage{
CalcClusterAverages_Gruffi(
  col_name = "Score.GO.0006096",
  obj = combined.obj,
  split_by = Seurat.utils::GetClusteringRuns(obj)[1],
  stat = c("mean", "median", "normalized.mean", "normalized.median")[3],
  scale.zscore = FALSE,
  quantile.thr = 0.99,
  absolute.thr = FALSE,
  plotit = FALSE,
  save.plot = FALSE,
  return.plot = FALSE,
  max.bin.plot = 50,
  simplify = TRUE,
  suffix = NULL,
  ylab.text = paste("Cluster", stat, "score"),
  title = paste("Cluster", stat, col_name),
  subtitle = NULL,
  width = 8,
  height = 6,
  xlb = if (absolute.thr) {
     paste("Threshold at", absolute.thr)
 } else {
    
    paste("Black lines: ",
    Stringendo::kppd(Stringendo::percentage_formatter(quantile.thr)), "quantiles |",
    "Cl. >", Stringendo::percentage_formatter(quantile.thr), "are highlighted. |",
    split_by)
 },
  ...
)
}
\arguments{
\item{col_name}{Name of the numeric metadata column to calculate averages for.
Default: 'Score.GO.0006096'.}

\item{obj}{Seurat single cell object, Default: combined.obj}

\item{split_by}{Name of the clustering (categorical metadata column) to calculate averages within.
Default: First clustering result from \code{Seurat.utils::GetClusteringRuns(obj)}.}

\item{stat}{Method to calculate the central tendency. Options are "mean", "median",
"normalized.mean", and "normalized.median". Default: "normalized.mean".}

\item{scale.zscore}{Logical flag to scale the calculated scores as z-scores. Default: FALSE.}

\item{quantile.thr}{Quantile threshold for cutoff, Default: 0.99}

\item{absolute.thr}{Use an absolute threshold? Default: FALSE}

\item{plotit}{Draw a plot? Default: FALSE}

\item{save.plot}{Save plot into a file? Default: FALSE}

\item{max.bin.plot}{Maximum number of clusters for which a barplot is considered feasible,
otherwise a histogram is drawn. Default: 50.}

\item{suffix}{Suffix to append to the plot/file name. Default: NULL.}

\item{ylab.text}{Y-axis label. Default: paste("Cluster", stat, "score")}

\item{title}{Plot title, Default: paste("Cluster", stat, col_name)}

\item{subtitle}{Plot subtitle, Default: NULL}

\item{width}{Plot width, Default: 8}

\item{height}{Plot height, Default: 6}

\item{xlb}{X-axis label for the plot, with dynamic content based on \code{absolute.thr} and \code{quantile.thr}.
Default varies based on these parameters.}

\item{...}{Additional parameters passed to internally called functions.}
}
\description{
Calculates the central tendency (mean or median) of specified scores for clusters (granules)
within a Seurat object and optionally plots these averages.
}
\seealso{
\code{\link[Stringendo]{percentage_formatter}}, \code{\link[Stringendo]{iprint}}
\code{\link[dplyr]{select_all}}, \code{\link[dplyr]{group_by_all}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}
\code{\link[rlang]{sym}}
\code{\link[CodeAndRoll2]{sem}}, \code{\link[CodeAndRoll2]{sortbyitsnames}}
}
