% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{CalcStandDevSkewedDistr}
\alias{CalcStandDevSkewedDistr}
\title{Calculate Standard Deviation for Skewed Distributions}
\usage{
CalcStandDevSkewedDistr(x, mean_x = mean(x), plothist = FALSE, breaks = 50)
}
\arguments{
\item{x}{A numeric vector representing the distribution.}

\item{mean_x}{The central tendency of \code{x}, used to mirror the distribution.
Default is the mean of \code{x}.}

\item{plothist}{Logical indicating whether to display histograms of the original, mirrored,
and combined distributions. Default is \code{FALSE}.}

\item{breaks}{The number of bins to use in the histogram if plotted.
Default is 50.}
}
\value{
The estimated standard deviation for the skewed distribution.
}
\description{
Computes a modified standard deviation for skewed distributions by mirroring data points about the mean.
This approach aims to reflect the skewed part of the distribution to better understand its spread.
}
\examples{
set.seed(123)
skewed_data <- rexp(100, rate = 0.1)
# Calculate modified standard deviation without plotting
modified_sd <- stand_dev_skewed(skewed_data)

# Calculate and plot the distribution and its mirrored version
stand_dev_skewed(skewed_data, plothist = TRUE)

}
