% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{CalcTranscriptomePercentage}
\alias{CalcTranscriptomePercentage}
\title{CalcTranscriptomePercentage}
\usage{
CalcTranscriptomePercentage(
  obj = combined.obj,
  genes = genes.GO.0061621.can.glyc
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{genes}{Charcter vector of genes, Default: genes.GO.0061621.can.glyc}
}
\description{
Calculate the percentage of transcriptome, for a given gene set.
}
\seealso{
\code{\link[Matrix]{colSums}}
\code{\link[Seurat]{reexports}}
}
