% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{CalcTranscriptomePercentageGO}
\alias{CalcTranscriptomePercentageGO}
\title{CalcTranscriptomePercentageGO}
\usage{
CalcTranscriptomePercentageGO(obj = combined.obj, GO.score = "GO.0061621")
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{GO.score}{GO Term. String of form "GO:xxxxxxx", Default: 'GO.0061621'}
}
\description{
Calculate the percentage of transcriptome, for a given GO-term gene set already stored`in obj@misc$GO[\link{GO.score}].
}
\seealso{
\code{\link[Matrix]{colSums}}
\code{\link[Seurat]{reexports}}
}
