% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{CalculateAndPlotGoTermScores}
\alias{CalculateAndPlotGoTermScores}
\title{Calculate and Plot GO Term Scores}
\usage{
CalculateAndPlotGoTermScores(
  obj = combined.obj,
  GO = "GO:0009651",
  data.base.access = c("biomaRt", "AnnotationDbi")[1],
  mirror = NULL,
  get.go.terms.if.missing = TRUE,
  open.browser = FALSE,
  plot.each.gene = FALSE,
  desc = "",
  save.UMAP = TRUE,
  verbose = TRUE,
  return.plot = FALSE,
  overwrite.misc.GO_genes = FALSE,
  ...
)
}
\arguments{
\item{obj}{A Seurat single-cell object.}

\item{GO}{The GO term to be analyzed, with a default value of "GO:0009651".}

\item{data.base.access}{Specifies the method to access gene list databases, with "biomaRt" as the
default option and "AnnotationDbi" as an alternative.}

\item{mirror}{Specifies an Ensembl mirror to use if the default connection fails, with \code{NULL} as
the default indicating no mirror is specified.}

\item{get.go.terms.if.missing}{Retrieves GO terms if they are missing in the data.
Defaults to \code{TRUE}.}

\item{open.browser}{If \code{TRUE}, opens the corresponding GO term page in a web browser. Defaults to
\code{FALSE}.}

\item{plot.each.gene}{If \code{TRUE}, generates a UMAP plot for each gene's expression in addition to
the GO term scores. Defaults to \code{FALSE}.}

\item{desc}{A description for the GO term to be included on the plot, with an empty string as the
default.}

\item{save.UMAP}{If \code{TRUE}, the UMAP plot is saved to a file. Defaults to \code{TRUE}.}

\item{verbose}{If \code{TRUE}, enables the printing of detailed messages throughout the function
execution. Defaults to \code{TRUE}.}

\item{return.plot}{If \code{TRUE}, the function returns the plot object instead of the updated Seurat
object. Defaults to \code{FALSE}.}

\item{overwrite.misc.GO_genes}{Overwrite preexisting gene set in \code{misc$gruffi$GO[[make.names(GO)]]}.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
Depending on the \code{return.plot} parameter, this function returns either a plot object or
an updated Seurat object with new GO term scores added.
}
\description{
Automates the process of obtaining and calculating of Gene Ontology (GO) term-based
gene set scores and it's visualization through UMAP plots.
}
\seealso{
\code{\link[Seurat]{reexports}}
\code{\link[Stringendo]{iprint}}
}
