% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{CalculateMedianClusterSize}
\alias{CalculateMedianClusterSize}
\title{Calculate Median Cluster Size}
\usage{
CalculateMedianClusterSize(
  obj,
  assay = NULL,
  res = NULL,
  suffix = "_snn_res.",
  columnName = NULL,
  q = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{obj}{A Seurat object containing metadata with clustering information.}

\item{assay}{The name of the assay to which the resolution parameter is applied.}

\item{res}{The resolution level for which to calculate the median cluster size.}

\item{suffix}{String suffix for clustering name}

\item{columnName}{Directly provide a meta.data column name instead of parsing it.}

\item{q}{quantile. Default: 0.5, i.e: the median.}

\item{verbose}{Logical; if TRUE, prints the median cluster size to the console. Defaults to TRUE.}
}
\value{
The median size of the clusters at the specified resolution.
}
\description{
This function calculates and returns the median size of clusters for a given assay
and resolution in a Seurat object. It provides an option to display the median size via message.
}
\details{
The function generates a column name based on the assay and resolution parameters,
checks for the existence of this column in the object's metadata, calculates the median size
of the clusters, and optionally outputs this information as a message.
}
\examples{
# Assuming `seuratObj` is a Seurat object with appropriate metadata
medianSize <- CalculateMedianClusterSize(obj = seuratObj, assay = "RNA", res = 0.8)

}
