% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{clean.duplicate.scorenames}
\alias{clean.duplicate.scorenames}
\title{clean.duplicate.scorenames}
\usage{
clean.duplicate.scorenames(obj = obj)
}
\arguments{
\item{obj}{Seurat single cell object, Default: obj}
}
\description{
Remove duplicate scorenames from obj@mata.data.
}
\seealso{
\code{\link[CodeAndRoll2]{grepv}}
\code{\link[Stringendo]{iprint}}
}
