% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_and_evaluate_custom_score.R
\name{custom_score_evaluation}
\alias{custom_score_evaluation}
\title{custom_score_evaluation}
\usage{
custom_score_evaluation(
  obj = combined.obj,
  custom.score.name = "Score.heGENES",
  clustering = "integrated_snn_res.48.reassigned",
  assay = "RNA",
  stat.av = c("mean", "median", "normalized.mean", "normalized.median")[3]
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{custom.score.name}{Name of your custom gene set.}

\item{clustering}{Which clustering to use (from metadata)? Default: if (is.null(sum(grepl(".reassigned", GetClusteringRuns(obj))))) GetClusteringRuns(obj)\link{1} else GetClusteringRuns(obj)\link{grepl(".reassigned", GetClusteringRuns(obj))}}

\item{assay}{Which assay to use?, Default: 'RNA'}

\item{stat.av}{How to caluclate the central tendency? Default: c("mean", "median", "normalized.mean", "normalized.median")\link{3}}
}
\description{
Custom gene-set derived score evaluation for filtering.
}
\seealso{
\code{\link[Seurat]{reexports}}
\code{\link[Stringendo]{iprint}}
\code{\link[IRanges]{character(0)}}
}
