% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{CustomScoreEvaluation}
\alias{CustomScoreEvaluation}
\title{CustomScoreEvaluation}
\usage{
CustomScoreEvaluation(
  obj = combined.obj,
  custom.score.name = "Score.heGENES",
  clustering = getGruffiClusteringName(obj),
  assay = "RNA",
  stat.av = c("mean", "median", "normalized.mean", "normalized.median")[3],
  ...
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{custom.score.name}{Name of your custom gene set.}

\item{clustering}{Which clustering to use (from metadata)? Default: getGruffiClusteringName(obj)
e.g. "integrated_snn_res.48.reassigned"}

\item{assay}{Which assay to use?, Default: 'RNA'}

\item{stat.av}{How to caluclate the central tendency?
Default: c("mean", "median", "normalized.mean", "normalized.median")\link{3}}
}
\description{
Custom gene-set derived score evaluation for filtering.
}
\seealso{
\code{\link[Seurat]{reexports}}
\code{\link[Stringendo]{iprint}}
}
