% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{.convert.GO_term.2.score}
\alias{.convert.GO_term.2.score}
\title{.convert.GO_term.2.score}
\usage{
.convert.GO_term.2.score(GO_term = "GO:0006096")
}
\arguments{
\item{GO_term}{GO-term; Default: "GO:0006096"}
}
\value{
Score name, e.g.: "Score.GO.0006096"
}
\description{
Convert a string GO_term-name to Score-name.
}
