% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{.convert.score.2.GO_term}
\alias{.convert.score.2.GO_term}
\title{.convert.score.2.GO_term}
\usage{
.convert.score.2.GO_term(ScoreNames = "Score.GO.0006096")
}
\arguments{
\item{ScoreNames}{Vector of ScoreNames to convert to GO term, Default: "Score.GO.0006096"}
}
\value{
GO term name, e.g.: "GO:0006096"
}
\description{
Convert a string Score-name to GO_term-name.
}
