% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{.fix.metad.colname.rm.trailing.1}
\alias{.fix.metad.colname.rm.trailing.1}
\title{.fix.metad.colname.rm.trailing.1}
\usage{
.fix.metad.colname.rm.trailing.1(obj = obj, colname = ScoreName)
}
\arguments{
\item{obj}{Seurat single cell object, Default: obj}

\item{colname}{Metadata column name to fix, Default: ScoreName}
}
\description{
Fix malformed (suffxed) column names in obj@meta.data. When running AddGOScore(),
a '1' is added to the end of the column name. It is hereby removed.
}
\seealso{
\code{\link[Stringendo]{iprint}}
}
