% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{.parse.GO}
\alias{.parse.GO}
\title{Parse GO Term from Granule-score Name}
\usage{
.parse.GO(
  ident = "RNA_snn_res.6.reassigned_cl.av_GO:0006096",
  pattern = "_cl\\\\.av_|^Score.",
  ...
)
}
\arguments{
\item{ident}{The column name to be split, e.g., 'RNA_snn_res.6.reassigned_cl.av_GO:0006096'.}

\item{pattern}{The pattern to split by, default '\emph{cl\\.av}'.}

\item{...}{Additional arguments passed to \code{strsplit}.}
}
\value{
GO Term as string, e.g. "GO:0006096"
}
\description{
Splits a Granule-score column name into its clustering resolution and
GO term, and returns the 2nd part, the GO Term.
}
\examples{
.parse.GO(ident = "RNA_snn_res.6.reassigned_cl.av_GO:0006096")
}
