% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{FeaturePlotSaveCustomScore}
\alias{FeaturePlotSaveCustomScore}
\title{FeaturePlotSaveCustomScore}
\usage{
FeaturePlotSaveCustomScore(
  obj = combined.obj,
  genes = "",
  name_desc = NULL,
  h = 7,
  PNG = T,
  ...
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{genes}{Charcter vector of genes, Default: ""}

\item{name_desc}{Descriptive name to suffix plot and file name, Default: NULL}

\item{h}{height, Default: 7}

\item{PNG}{Save as .png? Default: T}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot and save a Seurat FeaturePlot with a custom score in the meta data.
}
\seealso{
\code{\link[Seurat]{FeaturePlot}}
\code{\link[ggplot2]{labs}}
\code{\link[MarkdownHelpers]{ww.FnP_parser}}
\code{\link[cowplot]{save_plot}}
}
