% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{FeaturePlotSaveGO}
\alias{FeaturePlotSaveGO}
\title{FeaturePlotSaveGO}
\usage{
FeaturePlotSaveGO(
  obj = combined.obj,
  GO.score = "Score.GO.0034976",
  name_desc = NULL,
  save.plot = TRUE,
  title_ = paste(GO.score, name_desc),
  h = 7,
  PNG = T,
  ...
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{GO.score}{Metadata column name for a GO-term scoreDefault: 'Score.GO.0034976'}

\item{name_desc}{Descriptive name to suffix plot and file name, Default: NULL}

\item{save.plot}{Save plot into a file. Default: TRUE}

\item{title_}{Plot title, Default: paste(GO.score, name_desc)}

\item{h}{Plot height, Default: 7}

\item{PNG}{Save as .png? Default: T}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot and save a Seurat FeaturePlot from a GO-score.
}
\seealso{
\code{\link[Seurat]{FeaturePlot}}
\code{\link[ggplot2]{labs}}
\code{\link[Stringendo]{ww.FnP_parser}}
\code{\link[cowplot]{save_plot}}
}
