% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{FilterStressedCells}
\alias{FilterStressedCells}
\title{FilterStressedCells}
\usage{
FilterStressedCells(
  obj = combined.obj,
  res = "integrated_snn_res.30",
  quantile.thr = 0.9,
  GOterms = c(`glycolytic process` = "GO:0006096", `response to endoplasmic reticulum
    stress` = "GO:0034976"),
  direction = "above",
  saveRDS = TRUE,
  saveRDS.Removed = FALSE,
  PlotExclusionByEachScore = TRUE,
  PlotSingleCellExclusion = TRUE,
  GranuleExclusionScatterPlot = TRUE
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{res}{Clustering resolution, Default: 'integrated_snn_res.30'}

\item{quantile.thr}{Quantile threshold to cutoff stressed cells, Default: 0.9}

\item{GOterms}{GO-terms to use for filtering, Default: c(\verb{glycolytic process} = "GO:0006096", \verb{response to endoplasmic reticulum stress} = "GO:0034976")}

\item{direction}{filtering direction, above or below, Default: 'above'}

\item{saveRDS}{save RDS file, Default: T}

\item{saveRDS.Removed}{PARAM_DESCRIPTION, Default: F}

\item{PlotExclusionByEachScore}{PARAM_DESCRIPTION, Default: T}

\item{PlotSingleCellExclusion}{PARAM_DESCRIPTION, Default: T}

\item{GranuleExclusionScatterPlot}{PARAM_DESCRIPTION, Default: T}
}
\description{
Filter stressed cells from a Seurat single cell object.
}
\seealso{
\code{\link[Stringendo]{iprint}}, \code{\link[Stringendo]{percentage_formatter}}
\code{\link[Seurat.utils]{calc.cluster.averages}}
\code{\link[MarkdownHelpers]{filter_HP}}, \code{\link[MarkdownHelpers]{llprint}}
\code{\link[dplyr]{reexports}}
}
