% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{fix.metad.Colname.rm.trailing.1}
\alias{fix.metad.Colname.rm.trailing.1}
\title{fix.metad.Colname.rm.trailing.1}
\usage{
fix.metad.Colname.rm.trailing.1(obj = obj, colname = ScoreName)
}
\arguments{
\item{obj}{Seurat single cell object, Default: obj}

\item{colname}{Metadata column name to fix, Default: ScoreName}
}
\description{
Fix malformed (suffxed) column names in obj@meta.data.
}
\seealso{
\code{\link[Stringendo]{iprint}}
}
