% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{GetAllGOTermNames}
\alias{GetAllGOTermNames}
\title{Show All GO Terms in an Object, and retrieve corresponding names using AnnotationDbi}
\usage{
GetAllGOTermNames(obj = combined.obj, return.obj = TRUE)
}
\arguments{
\item{obj}{A Seurat object containing metadata with GO term score columns, default: \code{combined.obj}.}

\item{return.obj}{Logical indicating whether to return the modified Seurat object.}
}
\value{
A modified Seurat object with GO terms annotated in \code{misc$gruffi$GO.Lookup} or
a vector of GO term names.
}
\description{
Retrieves and optionally annotates all GO term identifiers present in the
metadata of a Seurat object. It identifies columns starting with 'Score.GO',
extracts the corresponding GO terms, retrieves tge corresponding term-names using AnnotationDbi,
and optionally saving them back into the object.
}
\seealso{
\code{\link[CodeAndRoll2]{grepv}}
\code{\link[AnnotationDbi]{GOID}}, \code{\link[AnnotationDbi]{GOTerms-class}}
}
