% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{GetAllGOTerms}
\alias{GetAllGOTerms}
\title{GetAllGOTerms}
\usage{
GetAllGOTerms(obj = combined.obj, return.obj = TRUE)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{return.obj}{Return Seurat object (or GO-names?) Default: TRUE, meaning object.}
}
\description{
Return all GO-terms present in a Seurat single cell object.
}
\seealso{
\code{\link[CodeAndRoll2]{grepv}}
\code{\link[AnnotationDbi]{GOID}}, \code{\link[AnnotationDbi]{GOTerms-class}}
}
