% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{GetClusteringRuns}
\alias{GetClusteringRuns}
\title{GetClusteringRuns}
\usage{
GetClusteringRuns(obj = combined.obj, res = FALSE, pat = "*snn_res.*[0-9]")
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{res}{Specific resolution to use? Default: FALSE}

\item{pat}{Search pattern to match in the name of the clustering run. Default: '\emph{snn_res.}\link{0-9}'}
}
\description{
Get clustering resolutions present in a Seurat single cell object.
}
\seealso{
\code{\link[CodeAndRoll2]{grepv}}
}
