% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{GetGruffiClusteringName}
\alias{GetGruffiClusteringName}
\title{Retrieve First Clustering Run or First Matching Pattern Run}
\usage{
GetGruffiClusteringName(
  obj,
  pattern = ".reassigned",
  granule.res.slot = "optimal.granule.res"
)
}
\arguments{
\item{obj}{A Seurat object with clustering results in meta.data.}

\item{pattern}{A character string pattern to filter clustering runs.
Default is ".reassigned".}
}
\value{
The name of the first clustering run that matches the pattern
if any; otherwise, the name of the first available clustering run.
}
\description{
Fetches the first clustering run from a Seurat object,
optionally filtered by a specific pattern (e.g., ".reassigned").
}
\examples{
# Assume `seuratObj` is your Seurat object
GetGruffiClusteringName(seuratObj)
GetGruffiClusteringName(seuratObj, pattern = "someOtherPattern")

}
