% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{GetNamedClusteringRuns}
\alias{GetNamedClusteringRuns}
\title{GetNamedClusteringRuns}
\usage{
GetNamedClusteringRuns(
  obj = combined.obj,
  res = c(F, 0.5)[1],
  topgene = F,
  pat = "^cl.names.Known.*[0,1]\\\\.[0-9]$"
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{res}{Clustering resolution, Default: c(F, 0.5)\link{1}}

\item{topgene}{Look for clustering run, where clusters are named by top gene. See github/vertesy/Seurat.pipeline, Default: F}

\item{pat}{Search pattern to match in the name of the clustering run. Default: '^cl.names.Known.*\link{0,1}\.\link{0-9}$'}
}
\description{
Get annotated clustering resolutions present in a Seurat single cell object.
}
\seealso{
\code{\link[CodeAndRoll2]{grepv}}
}
