% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{GOscoreEvaluation}
\alias{GOscoreEvaluation}
\title{GOscoreEvaluation}
\usage{
GOscoreEvaluation(
  obj = combined.obj,
  GO_term = "GO:0034976",
  new_GO_term_computation = FALSE,
  clustering = GetGruffiClusteringName(obj),
  save.UMAP = FALSE,
  plot.each.gene = FALSE,
  assay = "RNA",
  description = NULL,
  mirror = NULL,
  stat.av = c("mean", "median", "normalized.mean", "normalized.median")[3],
  ...
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{GO_term}{GO-term; Default: 'GO:0034976'}

\item{new_GO_term_computation}{Calculate new GO-term score? (or use exisitng one?) Default: FALSE}

\item{clustering}{Which clustering to use (from metadata)? Default: GetGruffiClusteringName(obj)
e.g. "integrated_snn_res.48.reassigned"}

\item{save.UMAP}{Save umap into a file. Default: FALSE}

\item{plot.each.gene}{Plot each gene's expression, Default: FALSE}

\item{assay}{Which assay to use?, Default: 'RNA'}

\item{description}{Description to added to plot title, e.g. GO-terms name, Default: 'NULL'}

\item{mirror}{Which Ensembl mirror to use in biomaRt::useEnsembl()? If connection to Ensembl fails with default settings, mirror can be specified. Default: 'NULL'}

\item{stat.av}{How to caluclate the central tendency? Default: c("mean", "median", "normalized.mean", "normalized.median")\link{3}}

\item{...}{Additional parameters to be passed to the function.}
}
\description{
GO-score evaluation for filtering.
}
\seealso{
\code{\link[Seurat]{reexports}}
\code{\link[Stringendo]{iprint}}
}
