% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{grScoreHistogram}
\alias{grScoreHistogram}
\title{Histogram Visualization of Granule Scores with Threshold}
\usage{
grScoreHistogram(
  obj = combined.obj,
  colname = i1,
  miscname = "thresh.stress.ident1",
  auto = TRUE,
  ...
)
}
\arguments{
\item{obj}{A Seurat object containing metadata with granule scores and miscellaneous information.
Defaults to \code{combined.obj}.}

\item{colname}{The name of the column within \code{obj@meta.data} containing granule scores.
Typically something like is \code{'RNA_snn_res.6.reassigned_cl.av_GO:0042063'}.}

\item{miscname}{The key within \code{obj@misc$gruffi} for retrieving the threshold value.
Default is \code{'thresh.stress.ident1'}.}

\item{auto}{A logical flag indicating whether to automatically invert the filtering logic
for a specific \code{miscname} condition. Default is TRUE.}

\item{...}{Additional parameters to be passed to \code{ggExpress::qhistogram}.}
}
\value{
The function directly calls \code{ggExpress::qhistogram} for plotting, hence does not return a value
but generates a histogram plot as a side effect.
}
\description{
Plots a histogram of granule scores from a specified column within a Seurat object,
with a vertical line denoting a threshold. The function allows for an optional
inversion of filtering logic based on the \code{miscname} and \code{auto} parameters.
}
\details{
The function first verifies the structure and content of the provided \code{obj}, ensuring it
contains the necessary components for threshold retrieval and score extraction. It then retrieves
the threshold and uses \code{CodeAndRoll2::as.numeric.wNames.character} to convert the specified granule
scores to numeric. The histogram is plotted using \code{ggExpress::qhistogram}, with customization options
available through additional arguments. The \code{auto} parameter's effect is noted, particularly for the
\code{thresh.notstress.ident3} condition, although it primarily impacts plotting logic interpretation.
}
\examples{
grScoreHistogram(
  obj = combined.obj,
  colname = "RNA_snn_res.6.reassigned_cl.av_GO:0042063",
  miscname = "thresh.stress.ident1"
)
}
