% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{grScoreUMAP}
\alias{grScoreUMAP}
\title{Wrapper for clUMAP with Thresholding and Plotting}
\usage{
grScoreUMAP(
  obj = combined.obj,
  colname = "RNA_snn_res.6.reassigned_cl.av_GO:0042063",
  miscname = "thresh.stress.ident1",
  auto = TRUE,
  ...
)
}
\arguments{
\item{obj}{An object containing the data and miscellaneous threshold information.
Default is \code{combined.obj}.}

\item{colname}{Name of the column containing granule scores.
Default is \code{'RNA_snn_res.6.reassigned_cl.av_GO:0042063'}.}

\item{miscname}{Name of the miscellaneous threshold information within \code{obj@misc$gruffi}.
Default is \code{'thresh.stress.ident1'}.}

\item{auto}{Automatically flip filtering}

\item{...}{Additional arguments to be passed to \code{clUMAP}.}
}
\description{
This function applies thresholding to granule scores within a specified object and column,
inverts the thresholding if a specific miscellaneous name is provided, and then calls clUMAP
for plotting based on the thresholded values.
}
\examples{
# Assuming combined.obj is available and properly formatted
customClUMAPWrapper(combined.obj,
                    'RNA_snn_res.6.reassigned_cl.av_GO:0042063',
                    'thresh.stress.ident1')

}
