% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{ParseGruffiGranuleScoreName}
\alias{ParseGruffiGranuleScoreName}
\title{Generate Gruffi Granule Score Name}
\usage{
ParseGruffiGranuleScoreName(
  goID,
  obj = combined.obj,
  granuleRes = GetGruffiClusteringName(obj)
)
}
\arguments{
\item{goID}{The Gene Ontology (GO) identifier.}

\item{obj}{A Seurat object with clustering results in meta.data.}

\item{granuleRes}{The granule resolution identifier.}
}
\value{
A character string representing the standardized score name.
}
\description{
This function creates a standardized score name using granule resolution and a
Gene Ontology (GO) identifier.
}
\examples{
GetGruffiGranuleScoreName(granuleRes = granule.res.4.gruffi, goID = "GO:0006096")
}
