% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{plot.clust.size.distr}
\alias{plot.clust.size.distr}
\title{plot.clust.size.distr}
\usage{
\method{plot}{clust.size.distr}(
  obj = combined.obj,
  category = if (is.null(sum(grepl(".reassigned",
    Seurat.utils::GetClusteringRuns(obj))))) Seurat.utils::GetClusteringRuns(obj)[1] else
    Seurat.utils::GetClusteringRuns(obj)[grepl(".reassigned",
    Seurat.utils::GetClusteringRuns(obj))],
  plot = T,
  thr.hist = 30,
  ...
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{category}{Clustering or any categorical metadata column name, Default: if (is.null(sum(grepl(".reassigned", Seurat.utils::GetClusteringRuns(obj))))) Seurat.utils::GetClusteringRuns(obj)\link{1} else Seurat.utils::GetClusteringRuns(obj)\link[=GetClusteringRuns(obj))]{grepl(".reassigned", Seurat.utils::GetClusteringRuns(obj))}}

\item{plot}{Plot results? Default: T}

\item{thr.hist}{Plot histogram or a barplot? If less than X categories, it draws a barplot. If more, a histogram, Default: 30}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot cluster size distribution.
}
\seealso{
\code{\link[ggExpress]{qbarplot}}, \code{\link[ggExpress]{qhistogram}}
}
