% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{plot_norm_and_skew}
\alias{plot_norm_and_skew}
\title{plot_norm_and_skew}
\usage{
plot_norm_and_skew(
  x,
  q,
  tresholding = c("fitted", "empirical")[1],
  plot.hist = TRUE,
  ...
)
}
\arguments{
\item{x}{Distribution}

\item{q}{Quantile}

\item{tresholding}{Tresholding calculation method, Default: c("fitted", "empirical")\link{1}}

\item{plot.hist}{Draw a histogram? Default: TRUE}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot normal distribution and skew.
}
