% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{PlotClustSizeDistr}
\alias{PlotClustSizeDistr}
\title{PlotClustSizeDistr}
\usage{
PlotClustSizeDistr(
  obj = combined.obj,
  category = GetGruffiClusteringName(obj),
  plot = TRUE,
  thr.hist = 30,
  ...
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{category}{Clustering or any categorical metadata column name.
Default: GetGruffiClusteringName(obj)}

\item{plot}{Plot results? Default: TRUE}

\item{thr.hist}{Plot histogram or a barplot? If less than X categories, it draws a barplot. If more, a histogram, Default: 30}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot cluster size distribution.
}
\seealso{
\code{\link[ggExpress]{qbarplot}}, \code{\link[ggExpress]{qhistogram}}
}
