% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{reassign.small.clusters}
\alias{reassign.small.clusters}
\title{reassign.small.clusters}
\usage{
reassign.small.clusters(
  obj = combined.obj,
  ident = obj@misc$gruffi$optimal.granule.res,
  cell.num = 30,
  reduction = c("pca", "3d_umap")[2],
  PCA_dim.recompution.umap.3d = ncol(obj@reductions$pca@cell.embeddings)
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{ident}{Identity (granules / clustering) to reassign, Default: obj@misc$gruffi$optimal.granule.res}

\item{cell.num}{Minimum number of cells per granule / cluster, Default: 30}

\item{reduction}{Dim. reduction used to estimate closest cluster center when reassigning. Default: c("pca", "3d_umap")\link{2}}

\item{PCA_dim.recompution.umap.3d}{If 3D umap need to be recomputed, how many PCA-dimensions to be used?, Default: ncol(obj@reductions$pca@cell.embeddings)}
}
\description{
Reassign granules (clusters) smaller than X to nearby granules based on 3D UMAP coordinates.
}
\seealso{
\code{\link[Stringendo]{iprint}}
\code{\link[Seurat]{RunUMAP}}, \code{\link[Seurat]{reexports}}
}
