% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{saveData}
\alias{saveData}
\title{saveData}
\usage{
saveData(new.thresh.glycolytic, new.thresh.ER.stress, save.dir = OutDir)
}
\arguments{
\item{new.thresh.glycolytic}{Threshold value for glycolysis}

\item{new.thresh.ER.stress}{Threshold value for ER.stress}

\item{save.dir}{Directory to save to, Default: OutDir}
}
\description{
Save data as RDS.
}
