% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{stand_dev_skewed}
\alias{stand_dev_skewed}
\title{stand_dev_skewed}
\usage{
stand_dev_skewed(x, mean_x = mean(x), plothist = F, breaks = 50)
}
\arguments{
\item{x}{distribution}

\item{mean_x}{central tendency, Default: mean(x)}

\item{plothist}{Show histogram? Default: F}

\item{breaks}{n.bins in histogram, Default: 50}
}
\description{
Standard deviation skew.
}
