% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{StressBarplotPerCluster}
\alias{StressBarplotPerCluster}
\title{Stress Barplot Per Cluster}
\usage{
StressBarplotPerCluster(
  obj = combined.obj,
  fill.by = "is.Stressed",
  group.by = GetClusteringRuns()[1],
  color_scale = rev((scales::hue_pal())(2)),
  custom_col_palette = TRUE,
  ...
)

StressBarplotPerCluster(
  obj = combined.obj,
  fill.by = "is.Stressed",
  group.by = GetClusteringRuns()[1],
  color_scale = rev((scales::hue_pal())(2)),
  custom_col_palette = TRUE,
  ...
)
}
\arguments{
\item{fill.by}{A character string indicating the column to use for fill. Default is 'is.Stressed'.}

\item{group.by}{A character string or a vector indicating the column(s) to use for grouping.
If NULL, uses the first clustering run.}

\item{color_scale}{A color scale to use for the plot. Defaults to the reverse of the hue palette.}

\item{custom_col_palette}{A logical value indicating whether to use a custom color palette. Default is TRUE.}
}
\value{
A bar plot visualizing the cell fractions per cluster, filled by the specified fill criteria.

A bar plot visualizing the cell fractions per cluster, filled by the specified fill criteria.
}
\description{
This function generates a bar plot for cell fractions, filled by stress status across clusters.
It utilizes the first clustering run for grouping and applies a custom color scale.

This function generates a bar plot for cell fractions, filled by stress status across clusters.
It utilizes the first clustering run for grouping and applies a custom color scale.
}
