% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StressUMAP.R, R/gruffi.R
\name{StressUMAP}
\alias{StressUMAP}
\title{Wrapper for clUMAP with Thresholding and Plotting}
\usage{
StressUMAP(
  obj = combined.obj,
  colname = "is.Stressed",
  GO.terms = names(obj@misc$gruffi$GO),
  title = paste("Stress identification"),
  suffix = NULL,
  ...
)

StressUMAP(
  obj = combined.obj,
  colname = "is.Stressed",
  GO.terms = names(obj@misc$gruffi$GO),
  title = paste("Stress identification"),
  suffix = NULL,
  ...
)
}
\arguments{
\item{obj}{An object containing the data and miscellaneous threshold information.
Default is \code{combined.obj}.}

\item{colname}{Name of the column containing granule scores.
Default is \code{'RNA_snn_res.6.reassigned_cl.av_GO:0042063'}.}

\item{GO.terms}{used for the computation}

\item{title}{Title of the plot. Default: "Stress identification".}

\item{suffix}{Suffix to append to the plot/file name. Default: NULL.}

\item{...}{Additional arguments to be passed to \code{clUMAP}.}

\item{miscname}{Name of the miscellaneous threshold information within \code{obj@misc$gruffi}.
Default is \code{'thresh.stress.ident1'}.}

\item{auto}{Automatically flip filtering}
}
\description{
This function applies thresholding to granule scores within a specified object and column,
inverts the thresholding if a specific miscellaneous name is provided, and then calls clUMAP
for plotting based on the thresholded values.

Plot Stress Assignment UMAP after running the gruffi pipeline
}
\examples{
# Assuming combined.obj is available and properly formatted
customClUMAPWrapper(
  combined.obj,
  "RNA_snn_res.6.reassigned_cl.av_GO:0042063",
  "thresh.stress.ident1"
)

StressUMAP(combined.obj)
}
