% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.optional.R
\name{UMAP3Dcubes}
\alias{UMAP3Dcubes}
\title{UMAP3Dcubes}
\usage{
UMAP3Dcubes(
  obj = combined.obj,
  n_bin = 10,
  plot = FALSE,
  save.plot = FALSE,
  reduction = c("umap", "pca", "tsne")[1],
  ident = Seurat.utils::GetClusteringRuns(obj)[1],
  PCA_dim = 50
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{n_bin}{Number of bins, Default: 10}

\item{plot}{Plot results? Default: FALSE}

\item{save.plot}{Save plot into a file. Default: FALSE}

\item{reduction}{Which dim. reduction to use? Default: c("umap", "pca", "tsne")\link{1}}

\item{ident}{Identity (metadata), Default: Seurat.utils::GetClusteringRuns(obj)\link{1}}

\item{PCA_dim}{PCA dimensions, Default: 50}
}
\description{
UMAP with 3D cubes.
}
\seealso{
\code{\link[Seurat]{RunUMAP}}, \code{\link[Seurat]{RunTSNE}}, \code{\link[Seurat]{reexports}}
\code{\link[sm]{binning}}
\code{\link[raster]{modal}}
\code{\link[viridis]{reexports}}
\code{\link[rgl]{open3d}}, \code{\link[rgl]{shade3d}}, \code{\link[rgl]{matrices}}, \code{\link[rgl]{cube3d}}, \code{\link[rgl]{rglwidget}}
\code{\link[htmlwidgets]{saveWidget}}
}
