% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{ww.convert.GO_term.2.score}
\alias{ww.convert.GO_term.2.score}
\title{ww.convert.GO_term.2.score}
\usage{
ww.convert.GO_term.2.score(GO_term = "GO:0006096")
}
\arguments{
\item{GO_term}{GO-term; Default: "GO:0006096"}
}
\description{
Convert a string GO_term-name to Score-name.
}
