% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{AddCustomScore}
\alias{AddCustomScore}
\title{AddCustomScore}
\usage{
AddCustomScore(
  obj = combined.obj,
  genes = "",
  assay.use = "RNA",
  FixName = TRUE
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{genes}{Charcter vector of genes, Default: ""}

\item{assay.use}{Which assay to use for getting the genes? Default: 'RNA'.}

\item{FixName}{Shall trailing "1" ar the end of the new column name be removed? Default: TRUE}
}
\description{
Add a custom gene expression score from a set of genes provided to the
Seurat single cell object. Calculates Score for gene set. Fixes name.
}
\seealso{
\code{\link[Seurat]{AddModuleScore}}
}
