% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{FilterStressedCells}
\alias{FilterStressedCells}
\title{Filter Stressed Cells from a Seurat Object}
\usage{
FilterStressedCells(
  obj = combined.obj,
  res = "integrated_snn_res.30",
  quantile.thr = 0.9,
  GOterms = c(`glycolytic process` = "GO:0006096",
    `response to endoplasmic reticulum stress` = "GO:0034976"),
  direction = "above",
  saveRDS = TRUE,
  saveRDS.Removed = FALSE,
  PlotExclusionByEachScore = TRUE,
  PlotSingleCellExclusion = TRUE,
  GranuleExclusionScatterPlot = TRUE
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{res}{Clustering resolution, Default: 'integrated_snn_res.30'}

\item{quantile.thr}{Quantile threshold to cutoff stressed cells, Default: 0.9}

\item{GOterms}{GO-terms to use for filtering, Default: c(\verb{glycolytic process} = "GO:0006096", \verb{response to endoplasmic reticulum stress} = "GO:0034976")}

\item{direction}{filtering direction, above or below, Default: 'above'}

\item{saveRDS}{Logical indicating if the filtered object should be saved as an RDS file, default is TRUE.}

\item{saveRDS.Removed}{Logical indicating if a separate RDS file for removed cells should be saved, default is FALSE.}

\item{PlotExclusionByEachScore}{Logical for plotting exclusion results by each score, default is TRUE.}

\item{PlotSingleCellExclusion}{Logical for plotting single cell exclusion results, default is TRUE.}

\item{GranuleExclusionScatterPlot}{Logical for plotting a scatter plot of granule exclusion, default is TRUE.}
}
\description{
Identifies and filters stressed cells based on specified GO terms and a quantile threshold.
It supports optional plotting of exclusion results and saving of modified datasets.
}
\seealso{
\code{\link[Stringendo]{iprint}}, \code{\link[Stringendo]{percentage_formatter}}
\code{\link[Seurat.utils]{calc.cluster.averages}}
\code{\link[MarkdownHelpers]{filter_HP}}, \code{\link[MarkdownHelpers]{llprint}}
\code{\link[dplyr]{reexports}}
}
