% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{GO_score_evaluation}
\alias{GO_score_evaluation}
\title{GO_score_evaluation}
\usage{
GO_score_evaluation(
  obj = combined.obj,
  GO_term = "GO:0034976",
  new_GO_term_computation = FALSE,
  save.UMAP = FALSE,
  plot.each.gene = FALSE,
  assay = "RNA",
  stat.av = c("mean", "median", "normalized.mean", "normalized.median")[3],
  clustering = if (is.null(sum(grepl(".reassigned", GetClusteringRuns(obj)))))
    GetClusteringRuns(obj)[1] else GetClusteringRuns(obj)[grepl(".reassigned",
    GetClusteringRuns(obj))]
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{GO_term}{GO-term; Default: 'GO:0034976'}

\item{new_GO_term_computation}{Calculate new GO-term score? (or use exisitng one?) Default: FALSE}

\item{save.UMAP}{Save umap into a file. Default: FALSE}

\item{plot.each.gene}{Plot each gene's expression, Default: FALSE}

\item{assay}{Which assay to use?, Default: 'RNA'}

\item{stat.av}{How to caluclate the central tendency? Default: c("mean", "median", "normalized.mean", "normalized.median")\link{3}}

\item{clustering}{Which clustering to use (from metadata)? Default: if (is.null(sum(grepl(".reassigned", GetClusteringRuns(obj))))) GetClusteringRuns(obj)\link{1} else GetClusteringRuns(obj)\link[=grepl(".reassigned",
GetClusteringRuns(obj))]{grepl(".reassigned", GetClusteringRuns(obj))}}
}
\description{
GO-score evaluation for filtering.
}
\seealso{
\code{\link[Seurat]{reexports}}
\code{\link[Stringendo]{iprint}}
\code{\link[IRanges]{character(0)}}
}
