% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{PlotGoTermScores}
\alias{PlotGoTermScores}
\title{PlotGoTermScores}
\usage{
PlotGoTermScores(
  obj = combined.obj,
  data.base.access = c("biomaRt", "AnnotationDbi")[1],
  mirror = NULL,
  only.draw.plot = FALSE,
  openBrowser = FALSE,
  plot.each.gene = FALSE,
  desc = "",
  save.UMAP = TRUE,
  verbose = TRUE,
  GO = "GO:0009651",
  ...
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{data.base.access}{Which tool to use to access gene list databases? Options:
biomaRt (deault) or AnnotationDbi.}

\item{mirror}{Ensembl mirror to use, helpful if the default connection fails with
default settings, mirror can be specified. Default: 'NULL'}

\item{only.draw.plot}{Only show GO term score umap plot, Default: FALSE}

\item{openBrowser}{Open website for GO term? Default: FALSE}

\item{plot.each.gene}{Plot each gene's expression on a combined umap? Default: FALSE}

\item{desc}{GO-score description on the plot, Default: ""}

\item{save.UMAP}{Save umap into a file. Default: TRUE}

\item{verbose}{Verbose output? Default: TRUE}

\item{GO}{GO-term; Default: 'GO:0009651'}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Automates retrieving, processing and plotting GO term based gene scores.
}
\seealso{
\code{\link[Seurat]{reexports}}
\code{\link[Stringendo]{iprint}}
}
