% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruffi.R
\name{PlotGoTermScores}
\alias{PlotGoTermScores}
\title{PlotGoTermScores}
\usage{
PlotGoTermScores(
  obj = combined.obj,
  use.ensemble = T,
  only.draw.plot = F,
  openBrowser = F,
  plot.each.gene = F,
  desc = "",
  save.UMAP = T,
  verbose = T,
  GO = "GO:0009651",
  ...
)
}
\arguments{
\item{obj}{Seurat single cell object, Default: combined.obj}

\item{use.ensemble}{Use ensemble to obtain scores? Default: T}

\item{only.draw.plot}{Only show GO term score umap plot, Default: F}

\item{openBrowser}{Open website for GO term? Default: F}

\item{plot.each.gene}{Plot each gene's expression on a combined umap? Default: F}

\item{desc}{GO-score description on the plot, Default: ""}

\item{save.UMAP}{Save umap into a file. Default: T}

\item{verbose}{Verbose output? Default: T}

\item{GO}{GO-term; Default: 'GO:0009651'}

\item{...}{Pass any other parameter to the internally called functions (most of them should work).}
}
\description{
Plot GO-term scores.
}
\seealso{
\code{\link[Seurat]{reexports}}
\code{\link[Stringendo]{iprint}}
}
